Added (size_t) cast for each %zu format to make sure that there are no issues on 64bit machines passing those numbers as 'int'

+++ Engine/ArchiveChannel.cpp
@@ -265,7 +265,7 @@ void ArchiveChannel::disable(Guard &guard, const epicsTime &when)
     if (disable_count > groups.size())
     {
         LOG_MSG("ERROR: Channel '%s' disabled %zu times?\n",
-                getName().c_str(), disable_count);
+                getName().c_str(), (size_t)disable_count);
         return;
     }
     // That handled the bookkeeping of being disabled or not.
